% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_category.R
\name{get_category}
\alias{get_category}
\title{get_category}
\usage{
get_category(symbol)
}
\arguments{
\item{symbol}{String specifying the name (id) of the time series.}

\item{api_key}{character specyfing the API for FRED}
}
\description{
Get the categories for an economic data series in the FRED database.
}
\examples{
get_category("GDPC1")
get_category("FEDFUNDS")
get_category("T10Y2Y")
}
