% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(symbol)
}
\arguments{
\item{symbol}{String specifying the name (id) of the time series.}

\item{api_key}{character specyfing the API for FRED}
}
\description{
Get information about a time series in the FRED database.
}
\examples{
get_metadata("GDPC1")
get_metadata("FEDFUNDS")
get_metadata("T10Y2Y")
}
