% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release.R
\name{get_release}
\alias{get_release}
\title{get_release}
\usage{
get_release(symbol)
}
\arguments{
\item{symbol}{String specifying the name (id) of the time series.}
}
\description{
Get release information for an economic data series in the FRED database.
}
\examples{
get_release("GDPC1")
get_release("FEDFUNDS")
get_release("T10Y2Y")
}
