% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source.R
\name{get_source}
\alias{get_source}
\title{get_source}
\usage{
get_source(source_id)
}
\arguments{
\item{source_id}{an ID, e.g. 1 for "Board of Governors of the Federal Reserve System (US)"}
}
\description{
Get the source of a specific time series.
}
\examples{
get_source(1)
get_source(3)
get_source(4)
}
