% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbol.R
\name{get_symbol}
\alias{get_symbol}
\title{get_symbol}
\usage{
get_symbol(symbol)
}
\arguments{
\item{symbol, }{character denoting the symbol to download, e.g. "GDPC1" or "FEDFUNDS"}

\item{api_key, }{character specyfing the api key for FRED}
}
\description{
Download data from FRED.
}
\examples{
get_symbol("GDPC1")
get_symbol("T10Y2Y")
c("FEDFUNDS", "GDPC1") |> map(get_symbol)
}
