% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols.R
\name{get_symbols}
\alias{get_symbols}
\title{get_symbols}
\usage{
get_symbols(symbol)
}
\arguments{
\item{symbols}{, character denoting the symbols to download, e.g. "GDPC1" or "FEDFUNDS"}

\item{api_key}{character specyfing the API for FRED}
}
\description{
Download data from FRED.
}
\examples{
get_symbols("GDPC1")
get_symbols("T10Y2Y")
get_symbols("FEDFUNDS", "T10Y2Y")
}
