% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/press_release_link.R
\name{press_release_link}
\alias{press_release_link}
\title{press_release_link}
\usage{
press_release_link(symbol)
}
\arguments{
\item{symbol}{String specifying the name (id) of the time series.}
}
\description{
Get release information for an economic data series in the FRED database.
}
\examples{
press_release_link("GDPC1")
press_release_link("FEDFUNDS")
press_release_link("T10Y2Y")
}
