% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_symbol.R
\name{search_symbol}
\alias{search_symbol}
\title{search_symbol}
\usage{
search_symbol(search_text)
}
\arguments{
\item{search_text, }{character containing the words with which you expect to find the id of the variable.}

\item{api_key, }{character specyfing the api key for FRED}
}
\description{
Search for the symbol of an economic data series that matches the \code{search_text}.
}
\examples{
search_symbol("Exports Mexico")
search_symbol("GDP Germany")
search_symbol("GNP France")
}
