% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols.R
\name{get_symbols}
\alias{get_symbols}
\title{get_symbols}
\usage{
get_symbols(symbol)
}
\arguments{
\item{symbol}{symbol, e.g. "GDPC1" or "FEDFUNDS"}
}
\description{
Download data from FRED.
}
\examples{
get_symbols("GDPC1")
get_symbols("FEDFUNDS")
get_symbols("T10Y2Y")
}
