% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_symbol.R
\name{search_symbol}
\alias{search_symbol}
\title{search_symbol}
\usage{
search_symbol(search_text, must_contain)
}
\arguments{
\item{search_text}{A string containing the words with which you expect to find the id of the variable.}

\item{must_contain}{A string like "Germany" to explicitly search for a variable which contains "Germany" in its title. Currently allows one expression only.}

\item{api_key}{character specyfing the API for FRED}
}
\description{
Search for the symbol of an economic data series that matches the search text.
}
\examples{
search_symbol("Exports", "Mexico")
search_symbol("GDP", "Germany")
search_symbol("G", "France")
}
